Initial Commit for the experimental version of phpDocumentor
With this version I have tried to apply DDD practices in the architecture
of phpDocumentor and by that make the architecture more friendly to new
developers while at the same time using that to enable new features, such
as support for multiple versions and native support for generating reference
documentation.

The version in this commit is not functional by a long shot as a whole lot
of details need to be taken care of, features are missing but as an experiment
it has potential.