Update Application Layer
I have changed Somos to use League/Tactician as CommandBus instead of
SimpleBus. This enabled me to remove the Tactician code from phpDocumentor
and thus simplify the initialization.

Since the latest changes in Somos fix several issues that were blocking for
this project I have also been able to use Somos itself instead of my
custom instantiation of PHP-DI.

On hindsight the Generator class does not really make sense since the
responsibilities belong to the Application layer and we pulled these
into the Domain Layer. By creating a new Command to create the Docs
and retrieve those by version number from the Repository I have been
able to remove the Generator class again.

Lastly I have added a configuration file for PHPUnit and wrote a first
unit test that can serve as basis for all other tests. The prototype
if still under development and heavily in flux so we will have to see
how much we can unit test in this phase.