Made a start with the 'new' Documentation
For the documentation I re-use the existing documentation of
phpDocumentor 2 and adapt it to match the BC breaks that will be
introduced.

Especially the 'internals' documentation has received a large upgrade
(though still not done) as it details the aggregates and how classes
are divided.

The internals documentation is also the primary source for contributors
to read up on how the parts of phpDocumentor interact.