Renamed DocumentationItem to DocumentGroup
In the UBIQUITOUS LANGUAGE we discovered that in a set of Documentation
you don't have an item in your documentation (DocumentationItem) but
more a group of documents that form a logical whole, such as a Guide
or Api Reference.

At the same time I have refactored the Documentation class to have a
single collection of DocumentGroups so that in the future we may
extend this with more concepts if that is necessary.